+++ scratch.js
@@ -3,10 +3,8 @@
 import nlp from './src/two/lib.js'
 // nlp.verbose(true)
 
-let doc = nlp(`hit Jardas al Abid's`, { 'Jardas al Abid': 'Cool' }).debug()
-
-// console.log(doc.model._multiCache)
-console.log(Object.keys(doc.model))
+let doc = nlp(`jack is guarded woops`)
+let m = doc.match('is guarded foo?$').debug()
 
 /*
 ['', '#'],