+++ tests/three/misc/abbreviation.test.js
@@ -3,12 +3,12 @@ import nlp from '../_lib.js'
 const here = '[three/abbreviation] '
 
 test('abbreviations', function (t) {
-  let doc = nlp(`mr. and Mrs. Kelly live on Shoreditch st.`)
+  let doc = nlp(`mr. and Mrs. Kelly live on Shoreditch st. in Canada`)
   doc.abbreviations().stripPeriods()
-  t.equal(doc.text(), 'mr and Mrs Kelly live on Shoreditch st.', here + 'no-periods')
+  t.equal(doc.text(), 'mr and Mrs Kelly live on Shoreditch st in Canada', here + 'no-periods')
 
   doc.abbreviations().addPeriods().addPeriods()
-  t.equal(doc.text(), 'mr. and Mrs. Kelly live on Shoreditch st.', here + 'one-period')
+  t.equal(doc.text(), 'mr. and Mrs. Kelly live on Shoreditch st. in Canada', here + 'one-period')
 
   t.end()
 })